moved setPlaymode to its own trait
